Makefile: merge runserver_dev and runserver targets
Now there is only the runserver target:

    make runserver # development
    NODE_ENV=production make runserver #production mode

Also this patch takes care of setting a default SESSION_SECRET for development
mode. But if we launch the server in production mode without explicitly setting
SESSION_SECRET, it will raise an error. The intent is to prevent setting a dumb
secret without noticing it.